pyproject.toml
[tool.poetry]
name = "dnac-sidekick"
homepage = "https://github.com/dannywade/dnac-sidekick"
version = "0.1.2"
description = "A CLI app used to interact with Cisco DNA Center"
authors = ["Dan Wade <danny.wade35@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "dnac_sidekick"}]

[tool.poetry.dependencies]
python = "^3.8"
attrs = "22.1.0"
black = "22.8.0"
certifi = "2022.6.15.1"
charset-normalizer = "2.1.1"
click = "8.1.3"
commonmark = "0.9.1"
coverage = "6.4.4"
idna = "3.3"
iniconfig = "1.1.1"
mypy-extensions = "0.4.3"
packaging = "21.3"
pathspec = "0.10.1"
platformdirs = "2.5.2"
pluggy = "1.0.0"
py = "1.11.0"
Pygments = "2.13.0"
pyparsing = "3.0.9"
pytest = "7.1.3"
pytest-cov = "3.0.0"
python-dotenv = "0.21.0"
requests = "2.28.1"
rich = "12.5.1"
tomli = "2.0.1"
typing-extensions = "4.3.0"
urllib3 = "1.26.12"
Jinja2 = "^3.1.2"
PyYAML = "^6.0"
genie = "^22.10"
pyats = "^22.10"

[tool.poetry.scripts]
dnac-sidekick = "dnac_sidekick.cli:dnac_cli"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
